requirements.txt
# Bates Agent Enhanced Dependencies
pypdf>=4.0.0
google-adk>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0